requirements.txt
opencv-python-headless
numpy
streamlit
torch
torchvision
Pillow
grad-cam
torchvision
